+++ packages/atproto_core/lib/src/const.dart
@@ -19,3 +19,6 @@ const defaultService = 'bsky.social';
 
 /// `bsky.network`
 const defaultStreamService = 'bsky.network';
+
+/// `cardyb.bsky.app`
+const defaultLinkPreviewService = 'cardyb.bsky.app';