+++ packages/bluesky/lib/src/embed/embed_service.dart
@@ -1,0 +1,42 @@
+// Copyright 2023 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// 📦 Package imports:
+import 'package:atproto_core/atproto_core.dart' as core;
+
+// 🌎 Project imports:
+import '../entities/link_preview.dart';
+
+/// Returns link preview information for a given [url].
+Future<core.Response<LinkPreview>> findLinkPreview(
+  final Uri url, {
+  String service = core.defaultLinkPreviewService,
+  core.GetClient? mockedGetClient,
+}) async =>
+    await _$EmbedService(
+      service: service,
+      mockedGetClient: mockedGetClient,
+    ).findLinkPreview(url);
+
+sealed class _EmbedService {
+  Future<core.Response<LinkPreview>> findLinkPreview(final Uri url);
+}
+
+final class _$EmbedService extends core.BaseHttpService
+    implements _EmbedService {
+  _$EmbedService({
+    required super.service,
+    super.mockedGetClient,
+  });
+
+  @override
+  Future<core.Response<LinkPreview>> findLinkPreview(final Uri url) async =>
+      await super.get(
+        '/v1/extract',
+        parameters: {
+          'url': url.toString(),
+        },
+        to: LinkPreview.fromJson,
+      );
+}