+++ packages/bluesky/lib/src/entities/link_preview.g.dart
@@ -1,0 +1,31 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: non_constant_identifier_names
+
+part of 'link_preview.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_LinkPreview _$$_LinkPreviewFromJson(Map json) => $checkedCreate(
+      r'_$_LinkPreview',
+      json,
+      ($checkedConvert) {
+        final val = _$_LinkPreview(
+          url: $checkedConvert('url', (v) => v as String),
+          title: $checkedConvert('title', (v) => v as String),
+          description: $checkedConvert('description', (v) => v as String),
+          image: $checkedConvert('image', (v) => v as String),
+        );
+        return val;
+      },
+    );
+
+Map<String, dynamic> _$$_LinkPreviewToJson(_$_LinkPreview instance) =>
+    <String, dynamic>{
+      'url': instance.url,
+      'title': instance.title,
+      'description': instance.description,
+      'image': instance.image,
+    };