+++ packages/bluesky/lib/src/feeds/feeds_service.dart
@@ -2644,6 +2644,7 @@ final class _FeedsService extends BlueskyBaseService implements FeedsService {
           cursor: cursor,
         ),
         to: to,
+        authType: core.AuthType.authOptional,
       );
 
   core.Pagination<T> _paginateCustomFeed<T>({
@@ -2660,6 +2661,7 @@ final class _FeedsService extends BlueskyBaseService implements FeedsService {
           cursor: cursor,
         ),
         to: to,
+        authType: core.AuthType.authOptional,
       );
 
   Future<core.XRPCResponse<T>> _findFeedSkeleton<T>({