+++ packages/bluesky/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Release Note
 
+## v0.10.8
+
+- Added `uri` on `ListItem`. ([#1026](https://github.com/myConsciousness/atproto.dart/issues/1026))
+
+## v0.10.7
+
+- Removed `entities` field on `PostRecord`. Instead of this field, `entities` is merged internally to `facets` field.
+
+## v0.10.6
+
+- Added deprecated `entities` field on `PostRecord`. ([#1020](https://github.com/myConsciousness/atproto.dart/issues/1020))
+  - `LegacyFacet`
+  - `TextSlice`
+
+## v0.10.5
+
+- Added `.isReplyDisabled` and `.isNotReplyDisabled` properties on `PostViewer`. ([#1019](https://github.com/myConsciousness/atproto.dart/issues/1019))
+
+## v0.10.4
+
+- Added `.findLinkPreview` function. ThiIts returns preview information for a given URL. ([#1009](https://github.com/myConsciousness/atproto.dart/issues/1009))
+
+## v0.10.3
+
+- Fixed a bug that occurred in [#1010](https://github.com/myConsciousness/atproto.dart/issues/1010). getCustomFeed behaves differently depending on whether or not an access token is given. ([#1012](https://github.com/myConsciousness/atproto.dart/issues/1012))
+
+## v0.10.2
+
+- Changed auth type from access to anonymous for `app.bsky.feed.getFeed`. ([#1010](https://github.com/myConsciousness/atproto.dart/issues/1010))
+
 ## v0.10.1
 
 - Upgraded `atproto_core`. ([#999](https://github.com/myConsciousness/atproto.dart/issues/999))